+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ## [Unreleased]
 
+## [14.6.1]
+
+### Fixes
+
+- Fixed SIGINT [#77](https://github.com/rpcpool/yellowstone-jet/issues/77)
+
 ## [14.6.0]
 
 ### Features